# Improve performance of, and reduce allocations in, UIHintAttribute