--- a/src/libraries/System.ComponentModel.Annotations/src/System/ComponentModel/DataAnnotations/UIHintAttribute.cs
+++ b/src/libraries/System.ComponentModel.Annotations/src/System/ComponentModel/DataAnnotations/UIHintAttribute.cs
@@ -219,7 +219,6 @@ namespace System.ComponentModel.DataAnnotations
 
                 return true;
             }
-
         }
     }
 }